+++ lib/bcdd/result.rb
@@ -90,12 +90,24 @@ class BCDD::Result
     format('#<%<class_name>s type=%<type>p value=%<value>p>', class_name: self.class.name, type: type, value: value)
   end
 
+  def deconstruct
+    [type, value]
+  end
+
+  def deconstruct_keys(_keys)
+    { name => { type => value } }
+  end
+
   alias eql? ==
   alias data value
   alias data_or value_or
   alias on_type on
 
   private
+
+  def name
+    raise Error::NotImplemented
+  end
 
   def known(block)
     self.unknown = false