+++ test/bcdd/result_test.rb
@@ -52,6 +52,18 @@ module BCDD
       assert_raises(BCDD::Result::Error::NotImplemented) { result.value_or { 0 } }
     end
 
+    test '#deconstruct' do
+      result = Result.new(type: :ok, value: 1)
+
+      assert_equal([:ok, 1], result.deconstruct)
+    end
+
+    test '#deconstruct_keys' do
+      result = Result.new(type: :ok, value: 1)
+
+      assert_raises(BCDD::Result::Error::NotImplemented) { result.deconstruct_keys([]) }
+    end
+
     test '#==' do
       result = Result.new(type: :ok, value: 2)